FEATURE: Support more buttons per-category (#1)
This PR the functionality to add more that 1 button, and also provides the option to limit each button by category. Additionally, you can now change the button label and title for each button.

For example, you can have the following setup:

Button 1:

- Tags to add: Tag A, Tag B, Tag C
- In categories: ABC, DEF
- Auto close: Yes
- Close in how many days: 2
- Button label: Add A, B, C
- Button title: Add tags A, B and C

Button 2:

- Tags to add: Tag A, Tag D
- In categories: DEF, XYZ
- Auto close: No
- Button label: Add A, D
- Button title: Add tags A and D

This provides a much more flexible usage of this component.